StageManager::current_span is a value of an EMA

+++ src/assign/trail_saving.rs
@@ -54,8 +54,8 @@ impl TrailSavingIF for AssignStack {
         }
     }
     fn reuse_saved_trail(&mut self, cdb: &impl ClauseDBIF) -> PropagationResult {
-        let q = self.stage_scale.trailing_zeros() as u16
-            + (cdb.derefer(crate::cdb::property::Tf64::LiteralBlockEntanglement) as u16) / 2;
+        let q = self.stage_scale.trailing_zeros()
+            + (cdb.derefer(crate::cdb::property::Tf64::LiteralBlockEntanglement) as u32) / 2;
 
         #[cfg(feature = "chrono_BT")]
         let dl = self.decision_level();
